Autosetting dir, scheduled_at and status for build Hooks

+++ app/models/build.rb
@@ -5,13 +5,12 @@ class Build < ActiveRecord::Base
   STATUS_FAILED = "status_build_failed"
   belongs_to :project
   before_destroy :remove_build_dir
+  before_create :set_build_values
 
   def perform
     self.update_attributes!(:status => STATUS_PROGRESS)
     stdout = ""
-    project_dir = project.build_dir
     self.started_at = Time.now
-    self.build_dir = File.join(project_dir, self.scheduled_at.strftime("%Y%m%d%H%M%S") + "_" + commit[0, 10])
     stdout << Runner.execute("git clone #{project.vcs_source} #{self.build_dir} 2>&1")
     stdout << Runner.execute("cd #{self.build_dir} && #{project.task} 2>&1")
     status = $?.exitstatus
@@ -31,7 +30,13 @@ class Build < ActiveRecord::Base
 
   private
   def remove_build_dir
-    Rails.logger.debug("[BigTuna] Removing build dir #{self.build_dir}")
-    FileUtils.rm_rf(self.build_dir)
+    FileUtils.rm_rf(self.build_dir) if File.directory?(self.build_dir)
+  end
+
+  def set_build_values
+    project_dir = project.build_dir
+    self.build_dir = File.join(project_dir, self.scheduled_at.strftime("%Y%m%d%H%M%S") + "_" + commit[0, 10])
+    self.status = STATUS_IN_QUEUE
+    self.scheduled_at = Time.now
   end
 end
